+++ examples/editor/src/ModalsMap.jsx
@@ -1,7 +1,9 @@
 import { ModalsMap as VideoModalsMap } from 'wix-rich-content-plugin-video';
 import { ModalsMap as SoundCloudModalsMap } from 'wix-rich-content-plugin-sound-cloud';
+import { ModalsMap as GiphyModalsMap } from 'wix-rich-content-plugin-giphy';
 
 export default {
   ...VideoModalsMap,
   ...SoundCloudModalsMap,
+  ...GiphyModalsMap,
 };